+++ packages/widgetbook_cli/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- **FEAT**: Support semantics testing. ([#1809](https://github.com/widgetbook/widgetbook/pull/1809))
 - **FEAT**: Add `init` command. ([#1777](https://github.com/widgetbook/widgetbook/pull/1777))
 
 ## 4.0.0-alpha.1